backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
sentence-transformers
faiss-cpu
openai
websockets
streamlit